Move coordinator tests
