+++ tests/components/nice_go/snapshots/test_init.ambr
